fix: isolate order form logic inside, fix styles
