+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "books-store",
+  "name": "fake-store",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
