+++ src/constants/inputs/index.ts
@@ -1,0 +1,1 @@
+export const deniedNumbers = ['e', 'E', '+', '-'];
