+++ src/constants/orders/index.ts
@@ -13,4 +13,3 @@ export const initialOrder = {
   expiryDate: '',
   cvcCode: '',
 };
-export const deniedNumbers = ['e', 'E', '+', '-'];
